Replace libsawtooth state and database implementation with transact
Transact's state and database code can replace libsawtooth's
implementations. This is the first step in replace sawooth's
transaction execution platform with transact.

This commit replaces crate::database with transact::database, and
then deletes the local database module.

This commit also replaces the used of crate::state::MerkleDatabase
with transact::state::merkle::MerkleRadix::MerkleRadixTree.

The DecodedMerkleStateReader is also reimplemented to use MerkleRadixTree
from transact. This requires converting the StateDatabaseError implemented
in Transact into the local error (which required adding a few error types).
This was done in support of the state views.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>