Update the BlockValidator to use Transact's TEP
This commit update the BlockValidator and the ChainController to use
Transact's Executor and Schedulers instead of libsawtooths traits.
This will require changes to sawtooth-core FFI code.

To support this change, the multiple threads created have been reduced to
one thread. Also BlockValidator.validate_block() now sends the request to
the running thread and waits for a response instead of creating its own
validators for the block.

Also fixes a typo  BLOCKVALIDATION_QUEUE_RECV_TIMEOUT.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>