Only add valid batches to validation rule enforcer
Updates the `ValidationRuleEnforcer` to only store the info from batches
that are valid. This allows the calling code to reject these batches
because they are invalid, and to continue to use the validation rule
enforcer to attempt to add valid batches.

Also updates the `Rule::validate` method to take transaction info as an
`IntoIter<Item = &'a TxnInfo>` to be more flexible.

Signed-off-by: Logan Seeley <seeley@bitwise.io>