Implement GenesisController
The GenesisController will be used to create the initial block for
the first validator in the network. This block will have a previous
block id of NULL_BLOCK_IDENTIFIER and consensus data will only include
b"Genesis". The GenesisController to create the genesis block requires
that there is a genesis batch file that contains the initial state.
The genesis batch file must contain a batch that sets the
sawtooth.consensus.algorithm.version and sawtooth.consensus.algorithm.name.

This is a rewrite of the python version in sawtooth-validator. Instead of
implementing a GenesisConsensus, that does very little validation, the
GenesisController will handle setting the consensus data.

This commit also adds a GenesisControllerBuilder.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>