Replace signing module with cylinder
Replaces the `sawtooth::signing` module with the `cylinder` library,
which provides a similar interface and implementations.

The use of the `HashSigner` has been replaced by the secp256k1
implementation that is provided by cylinder; wherever the hash signer
was used, transact now creates a secp256k1 signer with a random key.

Signed-off-by: Logan Seeley <seeley@bitwise.io>