Implement block publisher
This is pure Rust rewrite of the block publisher from the Sawtooth
validator.

This publisher differs from the old publisher in a few ways:

- It's constructed using a builder instead of `new` and `start` methods
- It's shutdown using a `ShutdownSignaler` and the `await_shutdown`
method instead of a `stop` method
- The `on_chain_updated` method doesn't restart an in-progress block;
that choice is entirely up to consensus, and the publisher should not
assume that.
- Performs signer verification when a batch is scheduled, since it may
differ from when it was submitted based on the state of the previous
block.
- The `summarize` and `finalize` methods do not have a `force` argument,
which just restarted a block if it was empty; there is no need to do
this now, since the scheduler is not shutdown if the block is empty.
- The `summarize` and `finalize` methods ensure that all injected
batches are executed, which was not guaranteed previously.
- The publisher tracks committed transactions on its own, so it does not
need the `TransactionCommitCache`
- All errors are now propagated instead of panicking
- The publisher is now a single struct, and logic has moved almost
entirely from the candidate block into the publisher itself. The
candidate block is just a data structure.

Batches are also submitted differently for the new publisher

- The `IncomingBatchSender` does not check for duplicate IDs; this is
already handled by the completer.
- No `has_batch` method is provided by the sender, since it is not used.
- The submitter is notified when the queue is full; this can be checked
using the sender without acquiring a lock on the publisher.

Signed-off-by: Logan Seeley <seeley@bitwise.io>