Detect if a block is empty before publishing
Updates the block publisher to check if a block contains any valid
transactions before publishing a block. Prior to this change, if the
only transactions added to a block were invalid, the publisher would
publish the block anyway. This ensures that only a block with valid
transactions gets published.

This change includes a slight reorganization of where the candidate
block is created to support re-creating the candidate block when there
are no valid transactions on finalization.

Signed-off-by: Logan Seeley <seeley@bitwise.io>