Implement state_verifier in Rust
Verifies that the state root hash of all blocks is in state and if not,
reconstruct the missing state. Assumes that there are no "holes" in
state, ie starting from genesis, state is present for all blocks up to some
point and then not at all.

State verification should be run before starting up any other component to
make sure no other blocks are processed with incorrect state.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>