Update add_txn_receipt index assignment
Update the add_txn_receipt method to first retrieve the index of the
most recently added transaction receipt in the database if it exists
and then increment that value by 1 for each new receipt being added to
the database. This fixes the issue where previously uniqueness was not
guaraunteed for the index value.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>