Modify postgres and sqlite migrations
Modify the `transaction_receipt` tables in postgres/migrations and
sqlite/migrations to include a new nullable column 'service_id' this
column is meant to hold an ID that that is specific to a
`DieselReceiptStore` instance.

Additionally, remove the unique constraint from the idx column. This is
because indexes only need to be unique for each service id.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>