Add `service_id` to `ReceiptStoreOperations`
Add a `service_id` field to the `ReceiptStoreOperations` struct. The
`service_id` will be used to set the `service_id` field in the
`TransactionReceiptModel` when adding transaction receipts to the store
as well as filter results when retrieving and listing transaction
receipts inorder to scope the receipt store to a specific instance.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>