Add a database index on transaction_receipt
This change adds an index on the "idx" and "service_id" columns of the
transaction_receipt table.  These columns are frequently used for
queries.

Use of these columns  has been identified as the source of a performance
issue with the query to find the latest `idx` value for a given service
id (found at the start of the `AddTxnReceiptsOperation`).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>