Add PublishFactory and PublishHandle
The publisher is in charge of returning the next item to publish
to the Supervisor and consensus. It will return the Artifact to
agree upon, for example a block or a batch execution result.
The publisher needs to be able to continue to publish until it is
told to stop. The work may be canceled if the building Artifact
will no longer be valid, for example if the current chain head has changed.

The following design is done with the goal of defining Rust structs that
takes a set of traits. These traits would allow the same publisher structs
to be used in any situation (e.i. Scabbard or Sawtooth).

A supervisor will use a PublishFactory that will be used to start
the execution. Once started, a PublishHandle is returned that is used
to finish or cancel the publishing.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>