Add a prototype test for PublishFactory and handle
This commit contains a test that verifies that the PublishFactory
is able to start publishing an artifact and return a handle
to interact with the publishing thread when provided a set of
implementation of the required traits. The trait implementations
are simple and meant only for testing.

The handle is then used to verify the expected artifact is returned
when calling finished and nothing is updated when cancel is called.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>