Remove transact dependency
This change removes the transact dependency and updates the code to use
the internal creates.

It requires the removal of duplicate trait implementations for proto ->
protocol conversions. This is done through the removal of the private
modules entirely.

It also requires the addition of two protocol features as default
features: protocol-block and protocol-genesis. These depend on transact
features that were formerly default.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>