Enable wasm contract building via compose
This change updates the dockerfiles for the sabre contract examples. It
adds REPO_VERSION to the .env file, bin/get_version and a VERSION file.
It adds docker-compose-installed.yaml, and copy-debs.yaml to handle
building the two available contracts and copying out the resulting scar
files. Finally, it adds the ci-debs justfile recipe for the process.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>